test: split test files for data cleaning
