+++ tests/test_cleaning/test_jhu_data.py
@@ -1,56 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from pathlib import Path
 import warnings
 import pytest
 import pandas as pd
-from covsirphy import SubsetNotFoundError, PCRIncorrectPreconditionError
-from covsirphy import COVID19DataHub, DataLoader, LinelistData
-from covsirphy import Term, JHUData, CountryData, PopulationData
-from covsirphy import OxCGRTData, PCRData, VaccineData
+from covsirphy import SubsetNotFoundError, Term, JHUData
 from covsirphy import JHUDataComplementHandler
-
-
-class TestCOVID19DataHub(object):
-    def test_covid19dh(self):
-        with pytest.raises(TypeError):
-            COVID19DataHub(filename=None)
-        data_hub = COVID19DataHub(
-            filename=Path("input").joinpath("covid19dh.csv"))
-        # Citation (with downloading), disabled to avoid downloading many times
-        # assert isinstance(data_hub.primary, str)
-        # Retrieve the dataset from the server
-        data_hub.load(name="jhu", force=False, verbose=False)
-        with pytest.raises(KeyError):
-            data_hub.load(name="unknown")
-        # Citation (without downloading)
-        assert isinstance(data_hub.primary, str)
-
-
-class TestDataLoader(object):
-    def test_start(self):
-        with pytest.raises(TypeError):
-            DataLoader(directory=0)
-
-    def test_dataloader(self, jhu_data, population_data, oxcgrt_data,
-                        japan_data, linelist_data, pcr_data, vaccine_data):
-        # List of primary sources of COVID-19 Data Hub
-        data_loader = DataLoader()
-        assert data_loader.covid19dh_citation
-        # Data loading
-        assert isinstance(jhu_data, JHUData)
-        assert isinstance(population_data, PopulationData)
-        assert isinstance(oxcgrt_data, OxCGRTData)
-        assert isinstance(japan_data, CountryData)
-        assert isinstance(linelist_data, LinelistData)
-        assert isinstance(pcr_data, PCRData)
-        assert isinstance(vaccine_data, VaccineData)
-        # Local file
-        data_loader.jhu(local_file="input/covid19dh.csv")
-        data_loader.population(local_file="input/covid19dh.csv")
-        data_loader.oxcgrt(local_file="input/covid19dh.csv")
-        data_loader.pcr(local_file="input/covid19dh.csv")
 
 
 class TestJHUData(object):
@@ -210,118 +165,3 @@ class TestJHUData(object):
         # Error handling
         with pytest.raises(SubsetNotFoundError):
             jhu_data.map(country="Greece")
-
-
-class TestPopulationData(object):
-    def test_cleaning(self, population_data):
-        df = population_data.cleaned()
-        column_set = set(Term.STR_COLUMNS) | set([Term.N, Term.ISO3])
-        assert set(df.columns) == column_set
-
-    def test_total(self, population_data):
-        assert isinstance(population_data.total(), int)
-
-    def test_to_dict(self, population_data):
-        assert isinstance(population_data.to_dict(country_level=True), dict)
-        assert isinstance(population_data.to_dict(country_level=False), dict)
-
-    def test_value(self, population_data):
-        assert isinstance(population_data.value("Japan"), int)
-        old_value = population_data.value("Japan", date="01Mar2020")
-        assert isinstance(old_value, int)
-        with pytest.raises(KeyError):
-            population_data.value("Japan", "01Jan1000")
-        population_data.value("UK")
-
-    def test_update(self):
-        population_data = PopulationData(filename=None)
-        population_data.update(1000, "Moon")
-        assert population_data.value("Moon") == 1000
-        population_data.update(2000, "Moon")
-        assert population_data.value("Moon") == 2000
-
-    def test_countries(self, population_data):
-        assert isinstance(population_data.countries(), list)
-
-    def test_map(self, population_data):
-        warnings.filterwarnings("ignore", category=UserWarning)
-        population_data.map(country=None)
-        population_data.map(country="Japan")
-        with pytest.raises(NotImplementedError):
-            population_data.map(variable="Feeling")
-
-
-class TestOxCGRTData(object):
-    def test_cleaning(self, oxcgrt_data):
-        df = oxcgrt_data.cleaned()
-        assert set(df.columns) == set(OxCGRTData.OXCGRT_COLS)
-
-    def test_subset(self, oxcgrt_data):
-        with pytest.raises(KeyError):
-            oxcgrt_data.subset("Moon")
-        df = oxcgrt_data.subset("JPN")
-        assert set(df.columns) == set(OxCGRTData.OXCGRT_COLS_WITHOUT_COUNTRY)
-
-    def test_total(self, oxcgrt_data):
-        with pytest.raises(NotImplementedError):
-            oxcgrt_data.total()
-
-    def test_map(self, oxcgrt_data):
-        warnings.filterwarnings("ignore", category=UserWarning)
-        oxcgrt_data.map(country=None)
-        with pytest.raises(NotImplementedError):
-            oxcgrt_data.map(country="Japan")
-
-
-class TestPCRData(object):
-    def test_cleaning(self, pcr_data):
-        df = pcr_data.cleaned()
-        assert set(df.columns) == set(PCRData.PCR_COLUMNS)
-
-    def test_from_dataframe(self, pcr_data):
-        df = pcr_data.cleaned()
-        assert isinstance(PCRData.from_dataframe(df), PCRData)
-
-    def test_use_ourworldindata(self, pcr_data):
-        pcr_data.use_ourworldindata(
-            filename="input/ourworldindata_pcr.csv")
-
-    @pytest.mark.parametrize("country", ["Japan"])
-    def test_subset(self, pcr_data, country):
-        with pytest.raises(SubsetNotFoundError):
-            pcr_data.subset(country, end_date="01Jan2000")
-        df = pcr_data.subset(country)
-        df = pcr_data.subset(country, end_date="01Jan2021")
-        assert set(df.columns) == set(PCRData.PCR_NLOC_COLUMNS)
-
-    @pytest.mark.parametrize("country", ["Greece"])
-    def test_subset_complement(self, pcr_data, country):
-        with pytest.raises(NotImplementedError):
-            pcr_data.subset_complement(country)
-
-    @pytest.mark.parametrize("country", ["Greece"])
-    def test_records(self, pcr_data, country):
-        with pytest.raises(SubsetNotFoundError):
-            pcr_data.records(country, end_date="01Jan2000")
-        df, _ = pcr_data.records(country)
-        assert set(df.columns) == set(PCRData.PCR_NLOC_COLUMNS)
-
-    @pytest.mark.parametrize("country", ["Greece", "Italy", "Sweden"])
-    def test_positive_rate(self, pcr_data, country):
-        warnings.simplefilter("ignore", category=UserWarning)
-        pcr_data.positive_rate(country, show_figure=True)
-        df = pcr_data.positive_rate(country, show_figure=False)
-        assert set(
-            [PCRData.T_DIFF, PCRData.C_DIFF, PCRData.PCR_RATE]).issubset(df.columns)
-
-    @pytest.mark.parametrize("country", ["China"])
-    def test_positive_rate_error(self, pcr_data, country):
-        with pytest.raises(PCRIncorrectPreconditionError):
-            pcr_data.positive_rate(country, show_figure=False)
-
-    def test_map(self, pcr_data):
-        warnings.filterwarnings("ignore", category=UserWarning)
-        pcr_data.map(country=None)
-        pcr_data.map(country="Japan")
-        with pytest.raises(NotImplementedError):
-            pcr_data.map(variable="Feeling")
